requirements.txt
antspyx
matplotlib
nibabel
numpy
openpyxl
pandas
scipy
scikit-learn
pydicom
pybids
tqdm
SimpleITK
dcm2bids
dcm2niix
sphinx
sphinx_rtd_theme